move processor to pipeline package
