+++ internal/processors/filter_by_set_test.go
@@ -26,7 +26,7 @@ import (
 func TestSetFilter(t *testing.T) {
 	type test struct {
 		token            *models.Token
-		setFilter        *filterBySet
+		setFilter        *FilterBySet
 		name             string
 		shouldBeFiltered bool
 	}
@@ -44,7 +44,7 @@ func TestSetFilter(t *testing.T) {
 			token: &models.Token{
 				Type: models.TokenTypeSIndex,
 			},
-			setFilter: &filterBySet{
+			setFilter: &FilterBySet{
 				setsToRestore: map[string]bool{
 					"test": true,
 				},
@@ -57,7 +57,7 @@ func TestSetFilter(t *testing.T) {
 				Type:   models.TokenTypeRecord,
 				Record: record,
 			},
-			setFilter:        &filterBySet{setsToRestore: map[string]bool{}},
+			setFilter:        &FilterBySet{setsToRestore: map[string]bool{}},
 			shouldBeFiltered: false,
 		},
 		{
@@ -66,7 +66,7 @@ func TestSetFilter(t *testing.T) {
 				Type:   models.TokenTypeRecord,
 				Record: record,
 			},
-			setFilter: &filterBySet{
+			setFilter: &FilterBySet{
 				setsToRestore: map[string]bool{
 					"anotherSet": true,
 				},
@@ -80,7 +80,7 @@ func TestSetFilter(t *testing.T) {
 				Type:   models.TokenTypeRecord,
 				Record: record,
 			},
-			setFilter: &filterBySet{
+			setFilter: &FilterBySet{
 				setsToRestore: map[string]bool{
 					setName: true,
 				},
